requirements.txt
fastapi
uvicorn
openai
fastapi-jwt-auth
pydantic
sqlalchemy
databases
alembic
python-dotenv
pyspellchecker
